Add plugin load path relative to robohydra.js path
Improve the chances of Robohydra finding its system-wide plugin
directory.

[Esteban] Edited to move the extra directory at the end, and drop
node_modules/robohydra/plugins, which is not needed anymore. The
rationale for the directory order is that the most "local" plugins
should have preference, and the plugins that come with the
distribution should only be used if there isn't any other source
for them.