Make setHostHeader true by default
It's by far the most common and useful setting. Although strictly
speaking this breaks compatibility with older RoboHydras, I don't
expect a lot of people to be bitten by this. On the other hand,
I'm *constantly* bitten by the fact that setHostHeader is false
by default. I don't want to imagine how much this default must
bring to other people, considering that debugging this is not the
most trivial thing, esp. if you're not even aware that there is an
option for it in the first place.