Add reset method to heads, call it on scenario start
When heads have state (eg. RoboHydraHeadStatic when used with
multiple responses) and they are included in a scenario, they
need a way to go back to their initial state when the scenario
is started (otherwise, starting the scenario again won't be
reliable as the state of the head on scenario start won't
always be the same).

To solve this problem, a reset() method is added. By default
it does nothing, but for RoboHydraHeadStatic it makes it
start with the first response again.