Keep compatibility with older, non-ES6-powered Nodes
The "qs" dependency had a wrong version specified, and it was a
simple ">=". Unfortunately, the latest version of "qs", as of two
days ago, uses ES6 features so it won't be compatible with older
Nodes. As there's nothing in RoboHydra that requires a recent
version of Node (it seems to work even in Node 0.6.x), force the
"qs" dependency to be slightly older for now. According to
https://github.com/hapijs/qs/issues/124#issuecomment-153220784 they
will provide security patches for the 0.5.x series, so we should be
fine.