Parse common content types into req.body
Instead of just having "rawBody" and "bodyParams" available, try
to parse the body into a string, object or whatever is appropriate
according to the content type and put the result in a new property
"body". If the content type is not supported or the body doesn't
have the correct format, "body" will be set to null.