Deprecate bodyParams
Now that we have req.body that contains the parameters when the
Content-Type is x-www-form-urlencoded, req.bodyParams is obsolete
and will be deprecated.

However, req.bodyParams works as long as the body contains the
right format, even if the Content-Type is wrong or missing. Hence,
we have to keep compatibility in the REST API by checking that the
Content-Type is correctly set in certain places, and warn the user
if it's not.

While at it, fix a bug with requests containing parameters in the
URL (like "/users/:id") and generated properties, like bodyParams
is now.