Drop websocket.io in favour of ws
Module websocket.io makes Node 6.1.0 spit out a warning when loaded
due to its use of process.EventEmitter. As RoboHydra was currently
using two WebSocket modules, and it seems that ws is able to do
what websocket.io was used for, the solution is to simply drop
websocket.io and only use ws for all WebSocket code.

Yay for dropping dependencies!