Add a pass-through option to RoboHydraFilesystem
Add an option to make RoboHydraFilesystem ignore the request (and
pass it through, as if it hadn't matched the RoboHydraFilesystem
head in the first place) if no matching static file is found.

This allows for custom 404 pages, easy file-based override of
content coming from a proxy, serving files from two possible
base directories, and other interesting combinations.